replace String#split with a custom parser
String#split removes nils (from unmatched captures), so we can't use it
to reliably reconstruct them. this was exposed by the backrefs test, which
is currently the only one that uses a pattern which contains optional
captures

Replacing String#split with a custom parser (based on String#scan) gives
us full access to the captures. it also simplifies the code since we no
longer need to work around the limitations and idiosyncrasies of the
String#split "protocol"

+ unskip the backrefs test